New sgfile parsing and execution (importing etc) now works (including param substitution).
Fixed the sgfile execution occasionally dropping changes -- was an issue with checkout not restarting
replication if it raised an exception (eg by not finding the image that it was asked to by the sgfile executor).
Suspending replication/recording changes etc is now a decorator to make sure we always restart it.
Did a pass with the pycharm formatter/import optimizer.