Change constants.py to pull values from CONFIG
Ultimately this is probably an unnecessary extra layer of indirection,
but this is the cleanest way to make the change right now, instead
of editing a bunch of files to pull from CONFIG.

In the next pass I think config should be passed as a click context
object. That should also happen around the time we refactor click
to do things like command grouping and context passing (e.g. to
share a single postgres connection).