Add ./clean_test.sh file for running tests from fresh state on dev machine
Run a "clean" test. Intended for running in local development to give a close
simulation of the test suite running in travis. Also accounts for the
possibility that some e.g. splitgraphci images are outdated.

  - Destroy and recreate docker environment (using down -v)
  - Run tests from within container
  - Exit 0 on all tests pass
  - Exit 1 otherwise

Give up on caching for now. Well, we can keep the pip cache, it
saves a couple seconds.