Include the parent ID in the hash that is used to calculate the new fragment ID.
This is for when the same action (with the same old and new rows) is performed on the two different versions of a table with the same schema: due to the way the object tree works, that pulls in all of the object's parents at materialization time (see test_diff_fragment_hashing_reused_twice), resulting in the table being stored incorrectly.

Whilst this might waste space, this shouldn't happen often. Alternatives would be:

 * including all of the fragments a table consists of in its metadata (as opposed to tips of object tree chains) -- would consume a lot of extra space for tables with a lot of changes
 * adding another indirection:
   * an object tree with nodes that point to actual objects (multiple nodes can point to the same object)
   * every node has a parent showing which node a given object should be applied to
   * tables now point to nodes in this tree rather than actual objects
   * this would be more difficult to maintain (extra work at object creation and materialization time) + overhead from another metadata table.