Use the "staging table" that we copy data into to chunk it instead of the source table to get the table schema.
This is because if the source table is temporary, it's not actually reflected in information_schema and so we can't
get its schema (returns an empty list), which affects object hashes and breaks range indexing.

Also change range indexing to use the TableColumn class members rather than tuple index ordinals.