Don't install pglast and disable SQL validation on Windows, since the query parsing library isn't available on this platform.
This means that SQL in Splitfiles and import statements will not be checked to make sure they don't reach beyond the target schema or call extra stored procedures. This is fine for Splitfiles/imports the user wrote themselves but means they have to double-check untrusted Splitfiles or the provenance of images that they're running `sgr rebuild` on.