Fix a bug where doing Ctrl+C during a long download/upload would roll back the whole push/pull routine and not register any objects. During a retried push, the client would waste time reuploading objects that it had already uploaded to S3 (but didn't register them) whereas during a retried pull, the client would redownload all objects (which would then collide with existing physical files which hadn't been correctly ticked off in the cache status table + mounted as CStore FDW tables).
Now, if a KeyboardInterrupt is caught in the download/upload handler (or it couldn't transfer some objects for another reason), it returns the list of objects that it has transferred and the actual exception inside a new `IncompleteObjectTransferError`. This error then gets caught by the `ObjectManager` in case of a pull or `_sync` in case of a push and successful objects are registered, then the routine reraises the internal exception.

Caveats: during a push, we now first register objects and only then register images/tables; this means that after a failed push, some orphaned objects will stay on the registry (but not invalid images) -- they will eventually be garbage collected. Also, after a failed pull, whilst handling the exception, the `ObjectManager` inspects the CStore tables rather than what the handler claimed it downloaded, so this should also work for non-KeyboardInterrupt handler terminations.