Delete `registry_meta` schema and API/commandline functionality related to "publishing" images.
Publishing was an extra step required after push to make the image visible in the catalog \+ precalculate things like a data preview and the image's provenance. The only information that we don't already have about the image is its README which can be uploaded via a separate channel: everything else can be recalculated by the catalog without a large resource cost. In addition, we want to decrease friction in publishing an image to Splitgraph.

This also mean that a lot of code related to handling `registry_meta` like RLS is gone (since for all other schemata it's been superseded by an explicit SQL API in `splitgraph_api.sql` -- the registry doesn't let users manipulate tables in splitgraph_meta directly).