Format SQL used in Splitfile `IMPORT`/`SQL` statements before hashing it for provenance purposes.
This is to store a prettified SQL in provenance as well as to get a canonical representation so that minor changes to SQL formatting don't force a rebuild. It's questionable whether this is the format we want to use for hashing (since it has way too many line breaks and indentations) or we should use an even more stripped (e.g. hash the AST somehow) implementation.

Also add a flag to `import_tables` to skip validating import SQL (Splitfile executor does it before passing queries to `import_tables`).