Fix mysterious `Missing objects: ()` error and fix leaving engine in an inconsistent state when downloading objects made with unsupported extensions.
If mounting the object fails, actually make sure to propagate the reason back to the user and delete the downloaded object (so that the downloader doesn't think the object already exists). This can be wasteful and in the future we should fix the logic to somehow take into account both objects that weren't downloaded and objects that didn't have a foreign table created for them but just asking the user to redownload is good enough for now.

Also, don't try to run various cache management functions etc if no objects were successful (cause of the `object_id IN ()` issue)