Fix issue with `manage_audit_triggers` looking at all repositories on the engine when the Repository object has a different object engine than the metadata engine)
`manage_audit_triggers` now doesn't create audit triggers in that case -- this means that repositories used in this manner won't delta compress but this is currently not used and wouldn't work anyway if the metadata engine is a registry where we can't write the HEAD tag.

A longer-term solution would be a special `checkout` table on the object engine where we can also indirect on repository <-> schema name (sidestepping issues with e.g. schemata with slashes not being supported by some tools).