Get the CSV plugin to accept old and new param schemas.
Instead of writing a config migration interface for data sources, punt on it
and get the CSV data source to attempt to migrate the parameters when it's
initialized (see if they have `url` or `s3_object, ...` at toplevel and push
them down into a separate `connection` object).

We still flatten these back when initializing the FDW, as PostgreSQL only lets us
pass a key-string value map (`OPTIONS ...`).