Fix a query performance issue with tables without a PK.
When partitioning data, if the table doesn't have a primary key, we use a "surrogate"
primary key by concatenating the whole row as a string on the PG side (this is because
the whole row can sometimes contain NULLs which we can't compare in PG).

The query planner used to use the non-textual composite PK to figure out which
chunks overlapped in order to materialize them. This meant false positives where
something like (2, 'orange') would force a materialization because numerically,
it's greater than 10 and lexicographically, it isn't.

To fix this, regenerate the textual surrogate PK at query plan time and use it
to figure out when chunks overlap.