Add experimental LQFDW that supports parallel scans and on-the-fly materialization.
Create a view that queries the Splitgraph image on the fly by reassembling objects.
It points to a union of "shim" foreign tables that determine whether or not they need
to be scanned through and, if so, downloading the backing object and running the scan.

This lets PG plan parallel scans through unions of CStore files and avoids the expensive
and sometimes buggy spillover of temporarily materialized tables to disk.