Add defaults to CSV params; bump default schema inference row size
The latter is a workaround for CU-1jad1fc but we can still hit it if e.g.
we have a column with 100,000 integers and a varchar after it. The only solution
in this case is to for the user to either manually change the column type or
to bump the inference size. We can detect this by scanning through the whole file
before we load it, but then that's redundant work.