Add functionality to the dbt utils module:
Add ability to compile and export the model's manifest file.

This sadly requires a connection to an engine, even though it just returns
a manifest.json file with compiled dbt views for every model. This is useful
for finding out which models a dbt repository provides as well as their
dependency tree.

Add ability to override each dbt data source's schema separately.

Add ability to filter a list of models when building a dbt model.