Serialize CSV foreign table/server options as JSON.
Instead of figuring out how to serialize/deserialize individual parameters as
strings, store each single one as JSON. We can't dump the whole params dict
to JSON as we rely on the Multicorn behaviour of table parameters overriding
server parameters.

This works around a bug where running a preview and passing a numerical parameter
(e.g. `autodetect_sample_size`) as a table parameter would make the FDW's
`IMPORT FOREIGN SCHEMA` implementation return an integer, which Multicorn wouldn't
be able to parse (since PostgreSQL doesn't support numerical options in FDWs).

This changes the FDW's interface (as it gives back and expects strings in sometimes
a different format, for example, the option `encoding 'utf-8'` is serialized
as `encoding '"utf-8"'`), but we don't use it directly. The data source end-to-end
tests aren't changed, so this fix doesn't break that abstraction.