Fix CSV type inference in `sgr csv import`
When using `sgr csv import`, we load data using `COPY FROM STDIN` which doesn't
let us treat empty strings as NULLs, whereas with the CSV FDW we can do that.
To get around this, we use the empty strings in the type inference in the former
case (so that integer columns with empty strings still end up as VARCHARs).