Use PG cursors + FETCH to partition tables on commit
Instead of doing a pre-flight query to first figure out partition boundaries,
then chunking the table by filtering on those boundaries (poor performance on
tables without PKs and foreign tables), use a PG server-side cursor and run
FETCHes to add data to CStore files.

Two changes to behaviour: not making objects in parallel and moving the object
hashing to be after we've grabbed the data from the cursor into a temporary
object (after which we actually rename it). The latter might actually speed
things up, since the hashing will be done against the CStore object which is
faster to read.

[still WIP, couple broken tests. Also needs some refactoring since
`create_base_fragment` is now a spaghettifest and I'm not sure if some of its
args are no longer used]