Fix FDW previews when column names have percentage signs
These column names issues when using argument interpolation, as psycopg2 treats
them as arguments and raises an IndexError (not enough arguments). Fix by
escaping them with double-percentage signs.

Note that this isn't fixed everywhere, just on the FDW data source preview code
path.

Also add a test using a CSV file with a percentage sign, as this is how this bug
was first discovered.