Fix slow startup times on OSX
On OSX, use PyInstaller's "single directory" mode instead of "single binary" and
distribute sgr as an archive. The problem is that the single binary itself
extracts executables that OSX needs to scan, which happens on every startup and
takes about 10s. This only happens once if using the single-directory mode.