Add support to `splitgraph.spec` for building multi-file executables
- Pass `--onedir` argument intead of `-F` when calling pyinstaller

Since spec file is just a Python file, add conditional logic
to use Python to create the appropriate `EXE()` and `COLLECT()`
TOCs in the spec file based on presence of `--onedir` flag.

Existing behavior is retained by default and will always build
a single-file executable, except when the `--onedir` flag is
contained in `sys.argv` of the `pyinstaller` command that is
processing the spec.