Update `install.sh` to default to multi-file on OS X [artifacts]
- Default to downloading release at `sgr-osx-x86_64.tgz`
when on OS X. This downloads a tarball of the multi-file
executable creatd with `pyinstaller --onedir`, which means
a directory that includes the `sgr` executable and its
dependencies. This performs better than executables generated
with `--onefile` on OS X.
- Allow overriding and reverting to previous behavior (i.e.,
downloading the single-file executable which is still addressable
at `sgr-osx-x86_64` in the release bundle) with `FORCE_ONEFILE` flag
- Unzip the package to `~/.splitgraph/pkg/sgr` (overwrite if exists),
then symlink `~/.splitgraph/sgr -> ~/.splitgraph/pkg/sgr/sgr`, in
order to keep compatibility with existing scripts
- Note: The self-upgrade (`sgr upgrade`) code might still need
to be changed.