Update OS X workflow to also build a multi-file executable
- Add steps to `osx_binary` stage to build multi-file executable
after also building the single-file executable, and also
upload that new artifact to the release

To support debugging, add pragma to workflow to build artifacts even if no tag

This will build all artifacts, but unless the commit is a tag,
it will not create a release or otherwise publish anything.

To enable this, also change `.ci/build_wheel.sh` to add support
for a `NO_PUBLISH` flag. When the `NO_PUBLISH` environment
variable is a non-empty value, do not error if PyPi
credentials are not included, and just run `poetry build`