Fix surrogate PK ordering
When performing LQ on a table without PKs, we need to ensure that the surrogate PKs that we use
(the string representation of the entire row) is properly ordered. Otherwise we will end up with some
objects being wrongly categorised as singletons as opposed to being grouped together, which will
lead to wrong query results.