Add explainer text to sgr engine add
As an "impatient onboarder" it wasn't immediately obvious to me the
implications of `sgr engine add` prompting for a password.
I first suspected `sgr` wanted root and believed it was a sudo prompt.

It seems reasonable to add some explainer text here, so as to assuage
any worries/confusion as to which password is expected/how it gets used.