Fix repeated rows when running sync with FDWs
Currently, we duplicate the previous image regardless of whether the ingestion
state exists and rely on the user having passed cursors if they call sync. If
they didn't, this results on the full data from the source table getting
appended to the target table again. To fix this, don't copy the tables from
the previous image if the user didn't pass any cursor fields.