Small hack to avoid flash-of-empty-image
When loading data, we create an image and sometimes can commit it to the DB
before the loading ends -- which makes that image latest. To work around this,
do a two-phase setup: first, set the image's creation time to 1BC, do the
ingestion and, if it succeeds, bump it to now() to make the image latest.